move the declarative task stuff out of the python backend testing
